(new) Nova estrutura CONTAS com Map para organizar os campos por grupos

+++ app/src/main/java/com/gussanxz/orgafacil/funcionalidades/contas/movimentacoes/visual/ui/ExibirItemListaMovimentacaoContas.java
@@ -1,19 +1,7 @@
 package com.gussanxz.orgafacil.funcionalidades.contas.movimentacoes.visual.ui;
 
-import com.gussanxz.orgafacil.funcionalidades.contas.movimentacoes.r_negocio.modelos.MovimentacaoModel;
+import com.gussanxz.orgafacil.funcionalidades.contas.movimentacoes.modelos.MovimentacaoModel;
 
-/**
- * MODEL: MovimentoItem
- * * RESPONSABILIDADE:
- * Atuar como um contêiner genérico (Wrapper) que permite ao RecyclerView carregar
- * múltiplos tipos de layouts na mesma lista.
- *
- * * O QUE ELA FAZ:
- * 1. Definição de Tipos: Categoriza o item como HEADER (Cabeçalho de data) ou MOVIMENTO (Linha financeira).
- * 2. Armazenamento Flexível: Guarda dados de resumo diário (título e saldo) ou o objeto completo de Movimentação.
- * 3. Factory Methods: Fornece métodos estáticos ('header' e 'linha') para criar instâncias de forma limpa e legível.
- * 4. Suporte ao Adapter: É a classe base que o 'MovimentosAgrupadosAdapter' utiliza para decidir qual XML inflar.
- */
 public class ExibirItemListaMovimentacaoContas {
 
     public static final int TYPE_HEADER = 0;
@@ -21,15 +9,16 @@ public class ExibirItemListaMovimentacaoContas {
 
     public int type;
 
-    // para header
-    public String data;        // "06/12/2025"
-    public String tituloDia;   // "Hoje - 06/12/2025" ou "Ontem - ..."
-    public double saldoDia;
+    // Campos do Header
+    public String data;
+    public String tituloDia;
+    public int saldoDia; // Mudei para int (centavos)
 
-    // para linha
+    // Campos do Item
     public MovimentacaoModel movimentacaoModel;
 
-    public static ExibirItemListaMovimentacaoContas header(String data, String tituloDia, double saldoDia) {
+    // Factory para Header
+    public static ExibirItemListaMovimentacaoContas header(String data, String tituloDia, int saldoDia) {
         ExibirItemListaMovimentacaoContas item = new ExibirItemListaMovimentacaoContas();
         item.type = TYPE_HEADER;
         item.data = data;
@@ -38,6 +27,7 @@ public class ExibirItemListaMovimentacaoContas {
         return item;
     }
 
+    // Factory para Linha
     public static ExibirItemListaMovimentacaoContas linha(MovimentacaoModel m) {
         ExibirItemListaMovimentacaoContas item = new ExibirItemListaMovimentacaoContas();
         item.type = TYPE_MOVIMENTO;
